requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
opencv-python==4.2.0.34
numpy==1.18.5
Werkzeug==1.0.1
opencv-contrib-python==4.5.4.58
pillow